test(orchestration): rename the corpus module to golden_tests
The user's S2 schema review picked golden vocabulary for the harness
test module. The file, the facade mod declaration, and the doc and
path references move; the committed snapshots are untouched because
their names derive from the normalize module path, not this one.

Card: S2
