+++ crates/aura/src/orchestration/context_fixture/golden_tests.rs
